+++ custom_components/ads_custom/switch.py
@@ -105,6 +105,7 @@ async def async_setup_entry(
                 hass,
                 device_id,
                 subentry.data.get(CONF_ENTITY_DEVICE_NAME) or name,
+                entry.entry_id,
             )
             device_identifiers = {(DOMAIN, device_id)}